+++ crates/stackable-operator/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Add `region` field to S3ConnectionSpec ([#959]).
+
+[#959]: https://github.com/stackabletech/operator-rs/pull/959
+
 ## [0.86.0] - 2025-01-30
 
 ### Added